Finish command
