+++ inscriptions/block-iframe-navigation.html
@@ -1,0 +1,3 @@
+<script>
+location.href = "https://google.com";
+</script>
