+++ inscriptions/content-security-policy.html
