+++ justfile
@@ -142,3 +142,6 @@ build-docs:
 
 update-changelog:
   git log --pretty='format:- %s' >> CHANGELOG.md
+
+preview:
+  cargo run preview inscriptions/*
